# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Novinzhstroy API
  description: |
    Comprehensive API for the Novinzhstroy project management system.

    ## Authentication
    This API uses JWT-based authentication with refresh tokens stored in HTTP-only cookies.

    ## Rate Limiting
    Authentication endpoints are rate-limited to prevent abuse.

    ## Versioning
    All endpoints are versioned under `/api/v1/` prefix.
  version: 1.0.0
  contact:
    name: API Support
    email: support@novinzhstroy.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.novinzhstroy.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login endpoint

    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HTTP-only refresh token cookie
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
          example: project_status_changed
        title:
          type: string
          example: Project status updated
        message:
          type: string
          example: Project "Office Ventilation" moved to review stage
        read:
          type: boolean
          example: false
        payload:
          type: object
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: List notifications
      description: Get user's notifications
      operationId: listNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
  /notifications/{notificationId}/read:
    post:
      tags:
        - Notifications
      summary: Mark notification as read
      description: Mark a specific notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
...